split rmap test
